refactor(core): move AssetsBaseUrlProvider from data-nats to openframe-core
Implementors of AssetsBaseUrlProvider are typically persistence-layer
modules, and openframe-data-nats carries io.nats:nats-spring-cloud-stream-
binder — depending on it for one interface leaked a second Spring Cloud
Stream binder onto consumers' classpaths, crashing services whose Kafka
bindings relied on single-binder mode (tenant-stream: "A component required
a bean of type Binder that could not be found").

The interface moves to com.openframe.core.service; DownloadConfiguration-
LinkResolver (data-nats) keeps consuming it via its existing openframe-core
dependency. Only known consumer of the old package is openframe-saas-lib
3.27.6, updated in lockstep.
