+++ openframe-core/src/main/java/com/openframe/core/service/.AssetsBaseUrlProvider.md
@@ -1,0 +1,18 @@
+<!-- source-hash: 904f373a44d087cdafecf6b498d37ee6 -->
+Strategy interface for supplying the base URL substituted for `{assetsBaseUrl}` in download link templates. This library ships **no implementation**: absent a bean, `DownloadConfigurationLinkResolver` (openframe-data-nats) falls back to the static `openframe.assets.base-url` property, so existing OSS and shared-plane deployments are unaffected.
+
+## Key Components
+
+| Member | Type | Description |
+|--------|------|-------------|
+| `getAssetsBaseUrl()` | Method | Returns the base URL (e.g. `https://tenant1.openframe.build`), or blank/null to defer to the property |
+
+## Why it lives in openframe-core
+
+Implementors are typically persistence-layer modules (the SaaS tenant plane derives the URL from the tenant's `tenants.domain` document). Placing the interface in `openframe-data-nats` — where its consumer, the link resolver, lives — forced those modules to depend on the NATS messaging module, whose pom carries `io.nats:nats-spring-cloud-stream-binder`. A second Spring Cloud Stream binder on the classpath breaks any service using another binder implicitly (tenant-stream's Kafka bindings crashed with "Binder bean not found"). `openframe-core` is dependency-light and already below both sides.
+
+## Contract
+
+- Called on every link resolution — implementations should be cheap (cache if backed by a datastore).
+- Returning blank or null is the "no opinion" signal, never an error: the resolver silently falls back to the property. Implementations must not throw for a missing value.
+- Trailing slashes are stripped by the resolver; implementations need not normalize.
