+++ openframe-core/src/main/java/com/openframe/core/service/AssetsBaseUrlProvider.java
@@ -1,11 +1,15 @@
-package com.openframe.data.nats.resolver;
+package com.openframe.core.service;
 
 /**
  * Supplies the base URL substituted for the {@code {assetsBaseUrl}} placeholder in download link
  * templates. No implementation is provided in this library: when no bean is present,
- * {@link DownloadConfigurationLinkResolver} falls back to the static
+ * {@code DownloadConfigurationLinkResolver} (openframe-data-nats) falls back to the static
  * {@code openframe.assets.base-url} property. Downstream deployments contribute a bean to resolve
  * the base URL dynamically (e.g. the SaaS tenant plane derives it from the tenant's domain).
+ *
+ * <p>Lives in openframe-core so implementors (typically persistence-layer modules) don't have to
+ * depend on the NATS messaging module — which carries a Spring Cloud Stream binder that must not
+ * leak onto the classpath of services using another binder.
  */
 public interface AssetsBaseUrlProvider {
