--- openframe-data-nats/src/main/java/com/openframe/data/nats/resolver/.AssetsBaseUrlProvider.md
@@ -1,16 +1,0 @@
-<!-- source-hash: 8817cd009a19c65b38193682728c7c1e -->
-Strategy interface for supplying the base URL substituted for `{assetsBaseUrl}` in download link templates. This library ships **no implementation**: absent a bean, `DownloadConfigurationLinkResolver` falls back to the static `openframe.assets.base-url` property, so existing OSS and shared-plane deployments are unaffected.
-
-## Key Components
-
-| Member | Type | Description |
-|--------|------|-------------|
-| `getAssetsBaseUrl()` | Method | Returns the base URL (e.g. `https://tenant1.openframe.build`), or blank/null to defer to the property |
-
-## Contract
-
-- Called on every link resolution — implementations should be cheap (cache if backed by a datastore).
-- Returning blank or null is the "no opinion" signal, never an error: the resolver silently falls back to the property. Implementations must not throw for a missing value.
-- Trailing slashes are stripped by the resolver; implementations need not normalize.
-
-Intended downstream implementation: the SaaS tenant plane derives the URL from the tenant's own `tenants.domain` document, so published download links carry the tenant subdomain and the shared gateway resolves the tenant's registered client version.
