+++ openframe-data-nats/src/main/java/com/openframe/data/nats/resolver/.DownloadConfigurationLinkResolver.md
@@ -1,4 +1,4 @@
-<!-- source-hash: a2dcea7958c72c551d2a51b6a89e4b79 -->
+<!-- source-hash: 094c87ee50d884a4107736a4a4997cbc -->
 Resolves download URLs from `DownloadConfiguration` link templates by substituting the `{version}` and `{assetsBaseUrl}` placeholders.
 
 ## Key Components
@@ -12,7 +12,7 @@ Resolves download URLs from `DownloadConfiguration` link templates by substituti
 
 ## Base URL Resolution
 
-The `AssetsBaseUrlProvider` constructor parameter is `@Nullable` — with the single constructor, Spring treats it as an optional dependency, so contexts without such a bean (OSS self-hosted, shared plane) fall back to the static property with no configuration change. The provider is consulted on every `resolve()` call, not captured at construction, so a value that changes at runtime (e.g. a tenant domain assigned after startup) is picked up. Trailing slashes are stripped from both sources before substitution.
+The `AssetsBaseUrlProvider` interface lives in `openframe-core` (`com.openframe.core.service`) so implementors don't need this module on their classpath. The constructor parameter is `@Nullable` — with the single constructor, Spring treats it as an optional dependency, so contexts without such a bean (OSS self-hosted, shared plane) fall back to the static property with no configuration change. The provider is consulted on every `resolve()` call, not captured at construction, so a value that changes at runtime (e.g. a tenant domain assigned after startup) is picked up. Trailing slashes are stripped from both sources before substitution.
 
 ## Usage Example
